backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.3
groq==0.11.0
openai==1.51.0
python-jobspy==1.1.75
pandas==2.2.3
aiohttp==3.10.9
fake-useragent==1.5.1
pdfplumber==0.11.4
python-multipart==0.0.9
